refactor: Exclui métodos não utilizados da classe FileDAO

+++ src/main/java/Views/EventoView.java
@@ -114,7 +114,7 @@ public class EventoView {
         listar();
 
         System.out.println("Escolha um ID:");
-        int id = ler.nextInt();
+        Long id = ler.nextLong();
         ler.nextLine();
 
         if (controller.deleteById(id)) {
@@ -136,7 +136,7 @@ public class EventoView {
         listar();
 
         System.out.println("Escolha um ID:");
-        int id = ler.nextInt();
+        Long id = ler.nextLong();
         ler.nextLine();
 
         Evento escolha = controller.getById(id);
